fix: node index starts from 1
