+++ utils/test-util.js
@@ -1,13 +1,26 @@
 // Whole-script strict mode syntax
 "use strict";
-const { toBN, UNITS, ERC1155 } = require("./contract-util");
+const {
+  toBN,
+  TOTAL,
+  MEMBER_COUNT,
+  UNITS,
+  ERC1155,
+} = require("./contract-util");
 const { toNumber } = require("web3-utils");
 const {
   expect,
   advanceTime,
   web3,
+  generateMembers,
   encodeProposalData,
 } = require("./hardhat-test-util");
+
+const {
+  createVote,
+  SigUtilSigner,
+  prepareVoteResult,
+} = require("./offchain-voting-util");
 
 const checkLastEvent = async (dao, testObject) => {
   let pastEvents = await dao.getPastEvents();
@@ -266,6 +279,258 @@ const lendERC1155NFT = async (
   expect(toNumber(unitBalance.toString())).to.be.closeTo(25000, 5);
 };
 
+/***
+ * Offchain Voting Test Utils
+ */
+
+const VotingStrategies = {
+  AllVotesYes: "AllVotesYes",
+  AllVotesNo: "AllVotesNo",
+  NoBodyVotes: "NoBodyVotes",
+  SingleYesVote: "SingleYesVote",
+  SingleNoVote: "SingleNoVote",
+  TieVote: "TieVote",
+  MajorityVotesYes: "MajorityVotesYes",
+  MajorityVotesNo: "MajorityVotesNo",
+};
+
+const testMembers = generateMembers(10);
+
+const findMember = (addr) =>
+  testMembers.find((member) => member.address === addr);
+
+const buildProposal = async (
+  dao,
+  actionId,
+  submitter,
+  blockNumber,
+  chainId,
+  name = "some proposal",
+  body = "this is my proposal",
+  space = "tribute"
+) => {
+  const proposalPayload = {
+    name,
+    body,
+    choices: ["yes", "no"],
+    start: Math.floor(new Date().getTime() / 1000),
+    end: Math.floor(new Date().getTime() / 1000) + 10000,
+    snapshot: blockNumber.toString(),
+  };
+
+  const startingTime = Math.floor(new Date().getTime() / 1000);
+  const proposalData = {
+    type: "proposal",
+    timestamp: startingTime,
+    space,
+    payload: proposalPayload,
+    submitter: submitter.address,
+  };
+
+  const signer = SigUtilSigner(submitter.privateKey);
+  proposalData.sig = await signer(proposalData, dao.address, actionId, chainId);
+  return { proposalData };
+};
+
+const vote = async (
+  dao,
+  bank,
+  proposalId,
+  member,
+  choice,
+  actionId,
+  chainId,
+  votingWeight = undefined,
+  signature = undefined
+) => {
+  const voteSigner = SigUtilSigner(member.privateKey);
+  const weight = await bank.balanceOf(member.address, UNITS);
+  const voteEntry = await createVote(
+    proposalId,
+    votingWeight ? toBN(votingWeight) : weight,
+    choice
+  );
+  voteEntry.sig = signature
+    ? signature
+    : voteSigner(voteEntry, dao.address, actionId, chainId);
+  return voteEntry;
+};
+
+const emptyVote = async (proposalId, choice = 2 /*no*/, signature = "0x") => {
+  const voteEntry = await createVote(proposalId, toBN("0"), choice);
+  voteEntry.sig = signature;
+  return voteEntry;
+};
+
+const buildVotes = async (
+  dao,
+  bank,
+  proposalId,
+  submitter,
+  blockNumber,
+  chainId,
+  actionId,
+  voteStrategy,
+  votingWeight = undefined,
+  voteChoice = undefined,
+  signature = undefined
+) => {
+  const membersCount = await bank.getPriorAmount(
+    TOTAL,
+    MEMBER_COUNT,
+    blockNumber
+  );
+  const voteEntries = [];
+  const totalVotes = parseInt(membersCount.toString());
+  for (let index = 0; index < totalVotes; index++) {
+    const memberAddress = await dao.getMemberAddress(index);
+    const member = findMember(memberAddress);
+    let voteEntry;
+
+    if (!member) {
+      voteEntry = await emptyVote(proposalId, 2, signature);
+      voteEntries.push(voteEntry);
+      continue;
+    }
+
+    switch (voteStrategy) {
+      case VotingStrategies.SingleYesVote &&
+        memberAddress === submitter.address:
+        voteEntry = await vote(
+          dao,
+          bank,
+          proposalId,
+          member,
+          voteChoice ? voteChoice : 1,
+          actionId,
+          chainId,
+          votingWeight,
+          signature
+        );
+        break;
+      case VotingStrategies.SingleNoVote && memberAddress === submitter.address:
+        voteEntry = await vote(
+          dao,
+          bank,
+          proposalId,
+          member,
+          voteChoice ? voteChoice : 2,
+          actionId,
+          chainId,
+          votingWeight,
+          signature
+        );
+        break;
+      case VotingStrategies.AllVotesYes:
+        voteEntry = await vote(
+          dao,
+          bank,
+          proposalId,
+          member,
+          voteChoice ? voteChoice : 1,
+          actionId,
+          chainId,
+          votingWeight,
+          signature
+        );
+        break;
+      case VotingStrategies.AllVotesNo:
+        voteEntry = await vote(
+          dao,
+          bank,
+          proposalId,
+          member,
+          voteChoice ? voteChoice : 2,
+          actionId,
+          chainId,
+          votingWeight,
+          signature
+        );
+        break;
+      case VotingStrategies.TieVote:
+        voteEntry = await vote(
+          dao,
+          bank,
+          proposalId,
+          member,
+          i < parseInt(totalVotes / 2) ? 1 : 2,
+          actionId,
+          chainId,
+          votingWeight,
+          signature
+        );
+        break;
+      case VotingStrategies.NoBodyVotes:
+      default:
+        voteEntry = await emptyVote(proposalId, 2, signature);
+        break;
+    }
+    voteEntries.push(voteEntry);
+  }
+  return voteEntries;
+};
+
+const buildVoteTree = async (
+  dao,
+  bank,
+  proposalId,
+  submitter,
+  blockNumber,
+  chainId,
+  actionId,
+  votingStrategy = VotingStrategies.AllVotesYes,
+  votes = undefined,
+  moveBlockTime = true
+) => {
+  const membersCount = await bank.getPriorAmount(
+    TOTAL,
+    MEMBER_COUNT,
+    blockNumber
+  );
+
+  const voteEntries = votes
+    ? votes
+    : await buildVotes(
+        dao,
+        bank,
+        proposalId,
+        submitter,
+        blockNumber,
+        chainId,
+        actionId,
+        votingStrategy
+      );
+
+  if (moveBlockTime) await advanceTime(10000);
+
+  const { voteResultTree, result } = await prepareVoteResult(
+    voteEntries,
+    dao,
+    actionId,
+    chainId
+  );
+
+  const signer = SigUtilSigner(submitter.privateKey);
+  const rootSig = signer(
+    { root: voteResultTree.getHexRoot(), type: "result" },
+    dao.address,
+    actionId,
+    chainId
+  );
+
+  const lastResult = result[result.length - 1];
+  lastResult.nbYes = lastResult.nbYes.toString();
+  lastResult.nbNo = lastResult.nbNo.toString();
+  return {
+    voteResultTree,
+    votesResults: result,
+    rootSig,
+    lastVoteResult: lastResult,
+    membersCount,
+    resultHash: voteResultTree.getHexRoot(),
+  };
+};
+
 module.exports = {
   checkLastEvent,
   checkBalance,
@@ -278,4 +543,12 @@ module.exports = {
   lendERC1155NFT,
   isMember,
   encodeDaoInfo,
+  testMembers,
+  VotingStrategies,
+  findMember,
+  buildProposal,
+  vote,
+  emptyVote,
+  buildVotes,
+  buildVoteTree,
 };
